requirements.txt
# Dependencias del sistema de Blockchain URP
mysql-connector-python>=8.0.0
